Distributed Student Performance Analytics System/Distributed Analytics System/requirements.txt
'''
COPY PASTE THIS IN TERMINAL FOR QUICK INSTALL (CTRL + `):

> all libraries:
    pip install -r requirements.txt

> only the essentials:
    pip install Flask Flask-Cors pymongo python-dotenv

'''

# Core Requirements
Flask==2.3.2
Flask-Cors==3.0.10
pymongo==4.5.0
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.0
pytest-cov==4.1.0
requests==2.31.0
Faker==19.6.2

# API Documentation (Optional)
flask-swagger-ui==4.11.1
flasgger==0.9.7.1

# Performance Optimization
gunicorn==21.2.0
gevent==23.9.1

# Data Analysis (Optional)
numpy==1.26.0
pandas==2.1.1

